fixed a few utils functions to get current day last entry, not just last entry in general
