+++ client/src/utils/dateFormat.js
@@ -27,9 +27,21 @@ function calculateAge(birthday) {
     return formattedDate;
   };
 
+  // get current date format just as it is saved in the backend (2-12-2024)
+  // // the above date formatter uses 2/12/2024 format
+  function getCurrentDate() {
+    const today = new Date();
+    const month = String(today.getMonth() + 1).padStart(2, '0');
+    const day = String(today.getDate()).padStart(2, '0');
+    const year = today.getFullYear();
+    
+    return `${month}-${day}-${year}`;
+  }
+
   module.exports ={
     calculateAge,
-    formatDate
+    formatDate,
+    getCurrentDate
   }
