+++ client/src/utils/editSeatChartUtil.js
@@ -1,13 +1,5 @@
 import { getBackgroundColorClass } from "./classroomColors";
-
-function getCurrentDate() {
-  const today = new Date();
-  const month = String(today.getMonth() + 1).padStart(2, '0');
-  const day = String(today.getDate()).padStart(2, '0');
-  const year = today.getFullYear();
-  
-  return `${month}-${day}-${year}`;
-}
+import { getCurrentDate } from "./dateFormat";
 
 export function getLastJournalInfo(student) {
   // returns if student has 0 journal entries
